decouple logger gui and runner
