+++ drive_characterization/__init__.py
@@ -7,7 +7,7 @@ import importlib.resources as resources
 def genRobotCode(projectType, config):
     if projectType == 'Simple':
         with resources.path(__name__, 'templates') as path:
-            with open(os.path.join(path, 'Simple', 'Robot.mako'), 'r') as template:
+            with open(os.path.join(path, 'Simple', 'Robot.java.mako'), 'r') as template:
                 return Template(template.read()).render(
                     diam=config['wheelDiameter'],
                     ppr=config['encoderPPR'],
@@ -25,7 +25,7 @@ def genRobotCode(projectType, config):
             )
     elif projectType == 'Talon':
         with resources.path(__name__, 'templates') as path:
-            with open(os.path.join(path, 'Talon', 'Robot.mako'), 'r') as template:
+            with open(os.path.join(path, 'Talon', 'Robot.java.mako'), 'r') as template:
                 return Template(template.read()).render(
                     diam=config['wheelDiameter'],
                     ppr=config['encoderPPR'],
@@ -42,5 +42,5 @@ def genRobotCode(projectType, config):
 
 def genBuildGradle(projectType, team):
     with resources.path(__name__, 'templates') as path:
-        with open(os.path.join(path, projectType, 'build.mako'), 'r') as template:
+        with open(os.path.join(path, projectType, 'build.gradle.mako'), 'r') as template:
             return Template(template.read()).render(team=team)
