+++ drive_characterization/templates/Simple/build.gradle.mako
